Language.Haskell.Tools.Demo

Documentation

type ClientId = Int

data RefactorSessionState

Constructors

RefactorSessionState

Fields

_refSessMods :: Map (String, String, FilePath) (UnnamedModule IdDom)

_actualMod :: Maybe (String, String, FilePath)

_isDisconnecting :: Bool

refSessMods :: Lens RefactorSessionState RefactorSessionState (Map (String, String, FilePath) (UnnamedModule IdDom)) (Map (String, String, FilePath) (UnnamedModule IdDom))

isDisconnecting :: Lens RefactorSessionState RefactorSessionState Bool Bool

actualMod :: Lens RefactorSessionState RefactorSessionState (Maybe (String, String, FilePath)) (Maybe (String, String, FilePath))

initSession :: RefactorSessionState

runFromCLI :: IO ()

runDemo :: [String] -> IO ()

app :: MVar [Int] -> FilePath -> Application

The application that is evoked for each incoming request

respondTo :: FilePath -> Int -> Session -> MVar RefactorSessionState -> (ByteString -> IO ()) -> ByteString -> IO ()

updateClient :: FilePath -> ClientMessage -> StateT RefactorSessionState Ghc (Maybe ResponseMsg)

This function does the real job of acting upon client messages in a stateful environment of a client

reloadAllMods :: FilePath -> StateT RefactorSessionState Ghc ()

createFileForModule :: FilePath -> String -> String -> IO ()

removeDirectoryIfPresent :: FilePath -> IO ()

moduleNameAndContent :: ((String, String, FilePath), mod) -> (SourceFileKey, mod)

dataDirs :: FilePath -> FilePath

userDir :: FilePath -> ClientId -> FilePath

initGhcSession :: FilePath -> IO Session

handleErrors :: FilePath -> ClientMessage -> (ResponseMsg -> IO ()) -> IO () -> IO ()

logToFile :: FilePath -> String -> ClientMessage -> IO ()

data ClientMessage

Constructors

KeepAlive

InitialProject

Fields

initialModules :: [(String, String)]

PerformRefactoring

Fields

refactoring :: String

moduleName :: String

editorSelection :: String

details :: [String]

ModuleChanged

Fields

moduleName :: String

newContent :: String

ModuleDeleted

Fields

moduleName :: String

Disconnect

data ResponseMsg

Constructors

RefactorChanges

Fields

moduleChanges :: [(String, Maybe String)]

ASTViewContent

Fields

astContent :: String

ErrorMessage

Fields

errorMsg :: String

CompilationProblem

Fields

errorMsg :: String

Disconnected